third_party/blink/web_tests/dom/mutation-event-tests/fast/dom/getElementById-consistency5-expected.txt

CONSOLE WARNING: Listener added for a 'DOMSubtreeModified' mutation event. This event type is deprecated, and will be removed from this browser VERY soon. Usage of this event listener will cause performance issues today, and represents a large risk of imminent site breakage. Consider using MutationObserver instead. See https://chromestatus.com/feature/5083947249172480 for more information.
Test that DOMSubtreeModified listeners cannot cause TreeOrderedMap to be incorrect

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS document.contains(el) is true
PASS document.getElementById('el') is null
PASS successfullyParsed is true

TEST COMPLETE
